refactor(map-calibration): move IconLikelihoodField into src/ Detection
Task 1 of the synthesis-J re-rank plan (docs/superpowers/plans/2026-06-02-synthesis-rerank.md).
Pure relocation: namespace Mithril.Tools.MapCalibrationFromScreenshot.SynthesisProbe ->
Mithril.MapCalibration.Detection; internal -> public. Body byte-identical.

Consumers updated with `using Mithril.MapCalibration.Detection;` where missing
(JEvaluator.cs, IconLikelihoodFieldSampleTests.cs). The other consumers
(SynthesisProbePhase, BuildTests, LoadDeviationTests, LoadDeviationRimMaskTests)
already had the using and so resolve to the moved type without edits.
