+++ tools/MapCalibrationFromScreenshot.SynthesisProbe.Tests/IconLikelihoodFieldSampleTests.cs
@@ -1,4 +1,5 @@
 using FluentAssertions;
+using Mithril.MapCalibration.Detection;
 using Mithril.Tools.MapCalibrationFromScreenshot.SynthesisProbe;
 using Xunit;
